+++ app/src/main/java/com/example/animemoi_app/common/navigation/AppNavigation.kt
@@ -110,7 +110,7 @@ fun AppNavigation() {
             composable(
                 route = "ComicDetailScreen",
             ) {
-                ComicDetailScreen()
+               // ComicDetailScreen()
             }
         }
     }
